# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AgriMarket South Sudan API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, assigned, in_transit, delivered, cancelled]
        - name: retailerId
          in: query
          schema:
            type: integer
        - name: truckId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order (retailer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status (admin/driver)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/assign-truck:
    patch:
      operationId: assignTruckToOrder
      tags: [orders]
      summary: Assign truck/driver to order (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TruckAssignment"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /analytics/demand:
    get:
      operationId: getDemandAnalytics
      tags: [analytics]
      summary: Product demand analytics over time
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Demand data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DemandDataPoint"
...
  /pricing:
    get:
      operationId: listPricing
      tags: [analytics]
      summary: Get current pricing rules (admin)
      responses:
        "200":
          description: Pricing rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PricingRule"
    post:
      operationId: createPricingRule
      tags: [analytics]
      summary: Create a pricing rule (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PricingRuleInput"
      responses:
        "201":
          description: Created pricing rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingRule"
components:
  schemas:
...
    LoginInput:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [name, phone, password, role]
      properties:
        name:
          type: string
        phone:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [farmer, retailer]
        farmName:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        language:
          type: string
          enum: [en, ar]
          default: en
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, phone, role, language, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, farmer, retailer]
        farmName:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        locationLat:
          type: ["number", "null"]
        locationLng:
          type: ["number", "null"]
        language:
          type: string
          enum: [en, ar]
        currency:
          type: string
          enum: [SSP, USD]
        isActive:
          type: boolean
        createdAt:
          type: string
...
    OrderItem:
      type: object
      required: [productId, quantity, priceSSP, priceUSD]
      properties:
        productId:
          type: integer
        productName:
          type: ["string", "null"]
        quantity:
          type: number
        unit:
          type: ["string", "null"]
        priceSSP:
          type: number
        priceUSD:
          type: number
    Order:
      type: object
      required: [id, retailerId, status, totalSSP, totalUSD, currency, items, createdAt]
      properties:
        id:
          type: integer
        retailerId:
          type: integer
        retailerName:
          type: ["string", "null"]
        truckId:
          type: ["integer", "null"]
        truckPlate:
          type: ["string", "null"]
        driverName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, assigned, in_transit, delivered, cancelled]
        totalSSP:
          type: number
        totalUSD:
          type: number
        currency:
          type: string
          enum: [SSP, USD]
        deliveryLocation:
          type: ["string", "null"]
        deliveryLat:
          type: ["number", "null"]
        deliveryLng:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
        updatedAt:
          type: string
    OrderInput:
      type: object
      required: [items, currency]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        currency:
          type: string
          enum: [SSP, USD]
        deliveryLocation:
          type: ["string", "null"]
        deliveryLat:
          type: ["number", "null"]
        deliveryLng:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
    OrderUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, assigned, in_transit, delivered, cancelled]
        notes:
          type: ["string", "null"]
    TruckAssignment:
      type: object
      required: [truckId]
      properties:
        truckId:
          type: integer
    OrderListResponse:
      type: object
      required: [orders, total]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
...
    DemandDataPoint:
      type: object
      required: [date, productName, orderCount, quantity]
      properties:
        date:
          type: string
        productName:
          type: string
        orderCount:
          type: integer
        quantity:
          type: number
...
    PricingRule:
      type: object
      required: [id, categoryId, minPriceSSP, maxPriceSSP, minPriceUSD, maxPriceUSD, updatedAt]
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        minPriceSSP:
          type: number
        maxPriceSSP:
          type: number
        minPriceUSD:
          type: number
        maxPriceUSD:
          type: number
        updatedAt:
          type: string
    PricingRuleInput:
      type: object
      required: [categoryId, minPriceSSP, maxPriceSSP, minPriceUSD, maxPriceUSD]
      properties:
        categoryId:
          type: integer
        minPriceSSP:
          type: number
        maxPriceSSP:
          type: number
        minPriceUSD:
          type: number
        maxPriceUSD:
          type: number