# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgriMarket South Sudan API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderCurrency } from "./orderCurrency";
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  retailerId: number;
  /** @nullable */
  retailerName?: string | null;
  /** @nullable */
  truckId?: number | null;
  /** @nullable */
  truckPlate?: string | null;
  /** @nullable */
  driverName?: string | null;
  status: OrderStatus;
  totalSSP: number;
  totalUSD: number;
  currency: OrderCurrency;
  /** @nullable */
  deliveryLocation?: string | null;
  /** @nullable */
  deliveryLat?: number | null;
  /** @nullable */
  deliveryLng?: number | null;
  /** @nullable */
  notes?: string | null;
  items: OrderItem[];
  createdAt: string;
  updatedAt?: string;
}