# lib/api-zod/src/generated/types/pricingRuleInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgriMarket South Sudan API
 * OpenAPI spec version: 0.1.0
 */

export interface PricingRuleInput {
  categoryId: number;
  minPriceSSP: number;
  maxPriceSSP: number;
  minPriceUSD: number;
  maxPriceUSD: number;
}