# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgriMarket South Sudan API
 * OpenAPI spec version: 0.1.0
 */
import type { UserCurrency } from "./userCurrency";
import type { UserLanguage } from "./userLanguage";
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  name: string;
  phone: string;
  role: UserRole;
  /** @nullable */
  farmName?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  locationLat?: number | null;
  /** @nullable */
  locationLng?: number | null;
  language: UserLanguage;
  currency?: UserCurrency;
  isActive?: boolean;
  createdAt: string;
}